+++ cpurast/framebuf.hpp
@@ -3,7 +3,6 @@
 #pragma once
 
 #include <vector>
-#include <cstddef>
 #include "color.hpp"
 
 using std::vector;
@@ -13,25 +12,25 @@ namespace cr
     class framebuf
     {
     public:
-        framebuf(size_t width, size_t height);
+        framebuf(unsigned int width, unsigned int height);
 
-        void resize(size_t new_width, size_t new_height);
+        void resize(unsigned int new_width, unsigned int new_height);
         void enable_depth_test();
         void clear(bool clear_color_buf, bool clear_depth_buf);
 
         // the origin of the coordinate system is placed at the left bottom
         // higher depth value means it is closer to the viewport
         // if depth test is disabled, depth value is ignored
-        void write(size_t x, size_t y, cr::color color, float depth);
+        void write(unsigned int x, unsigned int y, cr::color color, float depth);
 
-        size_t get_width() const;
-        size_t get_height() const;
+        unsigned int get_width() const;
+        unsigned int get_height() const;
         void set_clear_color(color clear_color);
         const vector<color>& get_color_buf() const;
 
     private:
-        size_t width;
-        size_t height;
+        unsigned int width;
+        unsigned int height;
 
         color clear_color; // color used for clearing color buffer
         vector<color> color_buf; // color buffer representation
